TS-OVER: move events out of User class
