+++ db.json
@@ -1,5 +1,10 @@
 {
   "users": [
-    
+    {
+      "id": 1
+    },
+    {
+      "id": 2
+    }
   ]
 }
