+++ src/Models/User.ts
@@ -1,4 +1,5 @@
 import axios, { AxiosResponse } from "axios";
+import { Eventing } from './Eventing';
 
 const host = 'http://localhost:3000/users/';
 
@@ -8,10 +9,9 @@ export interface UserProps {
     age?: number;
 };
 
-type Callback = () => void;
-
 export class User {
-    events: {[key: string]: Callback[]} = {};
+    public events: Eventing = new Eventing();
+
     constructor(private data: UserProps) {};
 
     get(propName: string): (number | string) {
@@ -21,18 +21,6 @@ export class User {
     set(update: UserProps): void {
         this.data = {...this.data, ...update}
     };
-
-    on(eventName: string, callback: Callback) {
-        const event = this.events[eventName] || [];
-        event.push(callback);
-        this.events[eventName] = event;
-    };
-
-    trigger(eventName: string): void {
-        const handler = this.events[eventName];
-        if(!handler || handler.length === 0) return;
-        handler.forEach(callback => callback());
-    }
 
     fetch(): void {
         axios.get(`${host}${this.get('id')}`)
